feat: twitter connect nextID (#5751)
* fix: disbale find name complate button when need verify next id

* refactor: extract collection verification method

* refactor: code style
