+++ packages/mask/src/social-network-adaptor/twitter.com/collecting/post.ts
@@ -1,25 +1,27 @@
-import { postsContentSelector, postsImageSelector } from '../utils/selector'
+import { postsContentSelector, postsImageSelector, timelinePostContentSelector } from '../utils/selector'
 import { IntervalWatcher, DOMProxy, DOMProxyEvents } from '@dimensiondev/holoflows-kit'
 import type { EventListener } from '@servie/events'
-import { creator, SocialNetworkUI as Next } from '../../../social-network'
+import { creator, globalUIState, SocialNetworkUI as Next } from '../../../social-network'
 import type { PostInfo } from '../../../social-network/PostInfo'
-import { postIdParser, postParser, postImagesParser } from '../utils/fetch'
+import { postIdParser, postParser, postImagesParser, postContentMessageParser } from '../utils/fetch'
 import { memoize, noop } from 'lodash-unified'
 import Services from '../../../extension/service'
 import { injectMaskIconToPostTwitter } from '../injection/MaskIcon'
-import { ProfileIdentifier } from '@masknet/shared-base'
+import { PostIdentifier, ProfileIdentifier } from '@masknet/shared-base'
 import {
     makeTypedMessageImage,
     makeTypedMessageTupleFromList,
     makeTypedMessageEmpty,
     makeTypedMessagePromise,
     makeTypedMessageTuple,
+    isTypedMessageText,
 } from '@masknet/typed-message'
 import { untilElementAvailable } from '../../../utils/dom'
 import { twitterBase } from '../base'
 import { twitterShared } from '../shared'
 import { createRefsForCreatePostContext } from '../../../social-network/utils/create-post-context'
 import { getCurrentIdentifier } from '../../utils'
+import { IdentityProviderTwitter } from './identity'
 
 function getPostActionsNode(postNode: HTMLElement | null) {
     if (!postNode) return null
@@ -142,6 +144,27 @@ export const PostProviderTwitter: Next.CollectingCapabilities.PostsProvider = {
     start(cancel) {
         registerPostCollectorInner(this.posts, cancel)
     },
+}
+
+export function collectVerificationPost(keyword: string) {
+    const userId = IdentityProviderTwitter.recognized.value.identifier || globalUIState.profiles.value[0].identifier
+    const postNodes = timelinePostContentSelector().evaluate()
+
+    for (const postNode of postNodes) {
+        const postId = postIdParser(postNode)
+        const postContent = postContentMessageParser(postNode)
+        const isVerified =
+            postId &&
+            postContent[0] &&
+            isTypedMessageText(postContent[0]) &&
+            (postContent[0].content ?? '').toLowerCase() === keyword.toLowerCase()
+
+        if (isVerified && userId) {
+            return new PostIdentifier(userId, postId)
+        }
+    }
+
+    return null
 }
 
 function collectPostInfo(
